+++ pkg/repository/match.go
@@ -159,7 +159,7 @@ type SatisfiedEntry struct {
 	DurableTaskInsertedAt pgtype.Timestamptz
 	NodeId                int64
 	Data                  []byte
-	InvocationCount       int64
+	InvocationCount       int32
 }
 
 type MatchRepository interface {
